[Kartik] Add .Added UC-2 in Address Book System Scenario Question

+++ oops-csharp-practice/scenario-based/AddressBookSystem/Contacts.cs
@@ -14,7 +14,12 @@ namespace BridgeLabzTraining.oops_csharp_practice.scenario_based.AddressBookSyst
         public string City { get; set; }
         public string State { get; set; }
         public int Zip {  get; set; }
-        public int PhoneNumber { get; set; }
+        public long PhoneNumber { get; set; }
         public string Email { get; set; }
+
+        //public override string ToString()
+        //{
+        //    return $"{FirstName} || {LastName} || {Address} || {City} || {State} || {Zip} || {PhoneNumber} || {Email}";
+        //}
     }
 }
